multibufworking

+++ libft/ft_putnbr_base.c
@@ -19,7 +19,7 @@ void		ft_putnbr_base(int nb, char *base)
 		ft_putchar('-');
 		nb = -nb;
 	}
-	if (nb >= ft_strlen(base))
+	if (nb >= (int)ft_strlen(base))
 	{
 		ft_putnbr_base(nb / ft_strlen(base), base);
 		ft_putnbr_base(nb % ft_strlen(base), base);
