+++ src/ft_handler_c.c
@@ -22,12 +22,12 @@ int		ft_handler_wc(t_flag *f, va_list *ap, t_list **begin_lst)
 	ws[1] = 0;
 	if (ws[0] < 0 && ws[0] > 55295 && ws[0] < 57344 && ws[0] > 1114111)
 		return (0);
-	f->size = ft_wcharlen(wc);
+	f->size = (size_t)ft_wcharlen(wc);
 	f->arg = ft_transform_wchar_in_char(ws);
 	f->free = 1;
 	if (f->flag[2] == 1 || f->flag[3] == 1 ||
 		f->flag[4] == 1 || f->flag[5] == 1 ||
-		f->flag[6] == 1 || f->flag[1] > f->size ||
+	    f->flag[6] == 1 || f->flag[1] > (int)f->size ||
 		f->flag[13] == 1)
 		return (ft_flags_char(f, begin_lst));
 	ft_multibuf_arg(f, begin_lst, f->size);
@@ -49,7 +49,7 @@ int		ft_handler_c(t_flag *f, va_list *ap, t_list **begin_lst)
 	f->size = 1;
 	if (f->flag[2] == 1 || f->flag[3] == 1 ||
 		f->flag[4] == 1 || f->flag[5] == 1 ||
-		f->flag[6] == 1 || f->flag[1] > f->size ||
+	    f->flag[6] == 1 || f->flag[1] > (int)f->size ||
 		f->flag[13] == 1)
 		return (ft_flags_char(f, begin_lst));
 	ft_multibuf_arg(f, begin_lst, f->size);
