+++ src/ft_handler_undefined.c
@@ -23,7 +23,7 @@ int		ft_handler_undefined(t_flag *f, t_list **begin_lst)
 	f->what = 1;
 	if (f->flag[2] == 1 || f->flag[3] == 1 ||
 		f->flag[4] == 1 || f->flag[5] == 1 ||
-		f->flag[6] == 1 || f->flag[1] > f->size ||
+	    f->flag[6] == 1 || f->flag[1] > (int)f->size ||
 		f->flag[13] == 1)
 		return (ft_flags_char(f, begin_lst));
 	ft_multibuf_arg(f, begin_lst, f->size);
