+++ src/ft_precision.c
@@ -14,10 +14,10 @@
 
 int		ft_precision(t_flag *f)
 {
-	if (f->flag[0] < f->size && f->flag[0] != -1)
+	if (f->flag[0] < (int)f->size && f->flag[0] != -1)
 		f->flag[13] = 0;
-	if (f->flag[0] >= f->size)
-		f->p = f->flag[0] - f->size;
+	if (f->flag[0] >= (int)f->size)
+		f->p = (size_t)f->flag[0] - f->size;
 	if (f->flag[2] == 1 && (f->c == 'o' || f->c == 'O'))
 		f->p -= f->x;
 	return (0);
